Linux: ptrace crash detection optimizations
* Add detection logic for multiple tid crashes against the
same target from same worker. Detection logic goes up
to depth 1. Proved useful for targets where multiple threads
process the same data for different reasons and have identical
crash backtraces from helper functions.

* Move crashes counter after initial checks that don't save
new crash triggers, to have an accurate value in display.

* Remove Android hack. Recently refactored ptrace
backend doesn't require manual signal jamming in
case of Android. ptrace logic ensures that target is
monitored at all time preventing debuggerd from
attaching to target (will fail at attach syscall). Also
previous bullet ensures that no duplicates are
analyzed for same master thread.

+++ fuzz.c
@@ -333,7 +333,7 @@ static void fuzz_fuzzLoop(honggfuzz_t * hfuzz)
         LOG_F("malloc(%zu) failed", hfuzz->maxFileSz);
     }
 
-    int rnd_index = util_rndGet(0, hfuzz->fileCnt - 1);
+    size_t rnd_index = util_rndGet(0, hfuzz->fileCnt - 1);
 
     /* If dry run mode, pick the next file and not a random one */
     if (hfuzz->flipRate == 0.0L && hfuzz->useVerifier) {
